Added the Oxygen Bar (Health Bar)
Added the Oxygen Bar on the left of the screen.  It goes red with Oxygen
is low and starts changing size rapidly as well.  Oxygen depletes at the
rate set in the PlayerHealth Script.  Right now there is a bug because
the tokens do not give you more oxygen as they should.  The bug is in
DemoCollectable.cs.   When the script PlayerHealth from Player is
accessed from DemoCollectable an this error happens:  object reference
not set to an instance of an object.